Update server.js
----------Commenting----------
Used the "Better Comments" extension in VS Code to help identify different page sections of my server.js
----------pageTitle----------
Added the variable pageTitle when rendering pages in order to have dynamic page titles
----------Admin Edit POST----------
Finished the app.post for event editing through the admin section. Correctly changed/updated date and name and redirects to admin to see changes in the events list there.
----------Admin Add----------
Completed functionality for the Admin Add page and POST in order to add a new event through user input
----------formatDate()----------
Revamped the formatDate() function to include a parameter and new formattedDate variable in order to correctly format the date for both registering as an attendee and event creation and editing
----------newId()----------
Created new function in order to generate a unique ID for both people registering and created events in order to ensure no events/people have the same ID, which would lead to issues with functionality and deleting attendees from events